folder_block_ops: avoid race that causes new block to be unref'd
If a new Write comes in right between the calls to `startSyncWrite`
and `makeLocalBcache`, a newly-made block could be added to the unrefs
list of the syncInfo, which could get it to be unref'd as part of the
same sync where it's being put to the server.  This causes the
DiskUsage of the resulting MD to be off by one block.

Instead, just use the syncInfo (and dirty directory entry) returned by
`startSyncWrite`, instead of the ones from the unrefCache and deCache,
which may have been modified by an intervening deferred write.

I also renamed `startSyncWrite` since it doesn't depend on the lock
being taken before it is called.