folder_branch_ops: don't block folder status on CR
If CR is stuck, we don't want to block the only way for users to
figure out if they're running CR.

However, we don't want to open ourselves up to races on the crChains
object, which could change after `setCRChains` is called.  So instead
just save the summary in the status object, instead of the full chain.
This doesn't cost us anything extra, since the summaries are already
being logged as part of every CR run.

Issue: KBFS-1315