dirty_bcache: reset sync buf capacity after 5 minutes
If no sync as large as the current buffer size was completed in the
last five minutes, this will reset the buffer size back to the minimum
and let it grow again according to the normal algorithm.  This lets us
adapt to suddenly changing network conditions (for example, a laptop
going offline and coming back online with a worse network).

Issue: KBFS-1414