kbfs_ops_test: fix ReIdentify test flake
This test tried to get around the fact that reidentifies were
asynchronous by sending twice on the reidentify channel.
Unfortunately, that led to a race, where identifyDone could be set
back to false by the second send, after the final identify was done.

Instead, the reidentify loop can explicitly signal when it's finished.

Issue: KBFS-1553