folder_block_ops: any file with a dirty deCache entry is dirty
If a file handle to a removed file gets one of its attributed set, it
will have a deCache entry even though the file itself might not be
dirty.  Then if a Sync happens, it won't count as dirty since no
blocks are dirty, and the sync won't clear out the deCache entry,
leading to a panic in the background flusher.

Issue: KBFS-1743