folder_block_manager: better check for MD equality
Two MDs may have the same data dir if they are both simple GCs, or one
is a GC and one is a no-op conflict resolution (say because the
unmerged branch contains only a GC).  So the code that checks whether
a canceled MD actually made it to the server should use an actual MD
ID for comparison, rather than the data dir.

Also, even if a GC op legitimately gets canceled and ends up on this
path, we should avoid ever archiving it, since that results in a
panic.

Issue: KBFS-1769