cr: don't let recoverable removal errors kill CR
If an unreferenced block can't be fetched, we shouldn't kill CR over
that.  It probably means a file/directory with broken/missing blocks
was removed from a TLF, and we should let it go through (with an
incorrect MD usage count for the TLF) just like we do for the
RemoveEntry() call.

Issue: KBFS-1761