folderlist: force a fast forward on a new user login
On a logout, we now reset the FBO TLF heads.  After the user re-logs
in, if FUSE still holds a reference to some node within the TLF, and
tries a read operation on it, it will fail with a "This request needs
MD write access, but doesn't have it.", because read operations aren't
normally allowed to set the head.

After a re-login, we want to invalidate any nodes that are still being
held, in the same way we do for fast-forwarding, and set the head, so
that the read operations don't need to.

Issue: KBFS-1923