block_journal: be sure to ignore all relevant MD markers
Recently we added an optimization in `ignoreBlocksAndMDRevMarkers`
where we'd stop iterating through the journal once we found all the
blocks we were looking for.  Unfortunately we missed an edge case: say
we're squashing revisions 5-15 together into one new revision, 5.  If,
while resolving the squash, we aren't ignoring any blocks from the
original revision 5, we will exit the loop early and never set the
`Ignored` flag on the original md marker entry for revision 5.  When
the flusher sees that unignored marker, it will think it's done with
the blocks for revision 5 early, and put the MD before all the blocks
have been flushed!  At that point, if the client crashes, or if
another device tries to read the folder, they could fail trying to
read a non-existant block.

Instead, don't exit early until we find the md marker for the earliest
revision in the squash.

Issue: KBFS-1939