folder_block_manager: push gcOps even when no blocks are deleted
This indicates to other devices (and future restarts of this device)
that there's no need to scan that MD range again looking for blocks to
delete.  This can happen for example when there's a range of gcOps
already at the tail of the TLF's history.

Also, we can avoid QR completely if the head consists of a gcOp that
deleted blocks all the way up to the head's parent revision.

Issue: KBFS-1922