file_data: set direct type correctly on new level of indirection
When making a new level of indirection, we were incorrectly setting
the DirectType of the first iptr of the new level -- it should have
the same type as the original top block, before the level was created.

In practice, this didn't cause any issues, because when the DirectType
should have been DirectBlock, it's always dirtied and the block type
recalculated anyway.  So this is just a cleanup to make things more
clear.

Issue: KBFS-1950