folder_branch_ops: don't start CR until the local head is set
Otherwise, if CR starts immediately and tries to look up the range of
unmerged updates, it won't include the (soon-to-be-set) head.  This
can cause CR to crash on an empty unmerged list, but even if we fix
that, it will be problematic because it might do CR over a partial set
of updates.

Issue: KBFS-1954