folder_branch_ops: don't panic if no sync progress is made
This can happen naturally if the journal is really full and flushing
slowly, and we don't want KBFS to completely break in that case.
Instead, print a warning and keep trying, though we should still wait
for the ticker so we don't end up in a tight loop for too long.