tlf_journal: avoid reading/signing an MD for no reason
When there are still blocks remaining to be flushed for the earliest
MD in the journal, we don't need to bother reading it from disk and
signing it, just to decide we aren't going to flush it after all.  Not
sure if this helps performance much, but at least it cuts down log
spam a bit.