folder_block_ops: don't zero out replaced block unrefs
si.unrefs holds blocks that need to be unreferenced as part of a sync.
If a block was created as part of a sync, and put to the bserver, but
the MD put failed and the block was then overwritten in the new MD
revision that does make it to the mdserver, we would "remove" it from
si.unrefs by zeroing out the pointer, but we'd keep around the block
size info so that it can be properly accounted for.

However, now the folderUpdatePrepper explicitly looks up the size of
each block, so that accounting is no longer needed, and we can
entirely remove the block.  Furthermore, the folderUpdatePrepper
doesn't like invalid block pointers, so this fixes an "Invalid block
ref" error after sync fails a few times.

Issue: KBFS-2157