folder_block_ops: can't cancel a write due to a sync context err
If a write is blocked on waiting for the dirty block cache to release
more bytes that can be dirtied, we have a mechanism where we can error
that write if something fatal and non-recoverable happens to an
existing sync that's currently draining the dirty bytes.  For example,
the user fills up their quota, and the sync fails, there's no reason
to accept the write knowing that it will just fail in the same way.

However, some tests cancels Syncs on purpose, but expect the next
Write to succeed, and there's no good way to tell for sure when the
Sync goroutine that was forked has finished.  (The actual Sync() call
could return earlier with the canceled error.)

We can safely ignore those errors, since the sync failing in that way
doesn't imply anything about syncs failing in the future.

Issue: KBFS-2164