kbfs: properly type TLFs, instead of using just a boolean
This ended up being a far-reaching change, not only because of
switching the TLF handles over, but because determining the proper
type of a handle for a RootMetadata object required storing the
Writers/Readers as keybase1.UserOrTeamIDs, instead of keybase1.UIDs.

This is a largely mechanical change that doesn't add much new logic.

Issue: KBFS-2185