libkbfs: check team members and resolve team IDs
To check writer/reader privileges (and MD signature validity) at the
level of the local servers, we need to pass contexts and a
TeamMembershipChecker to a lot more places.

Note that this isn't quite right, because it is using _current_ team
membership to check the validity of a signature on possibly-historic
MD updates.  Once we figure out the exact story with checking historic
team membership, this will need a bit of updating.

This makes a full libkbfs test of a team TLF possible!

Issue: KBFS-2192