folder_branch_ops: don't collect resolved paths for unlinked nodes
During SyncAll(), we don't want to bother adding resolved paths for
nodes that have already been unlinked, as this will result in the
folderUpdatePrepper attempting (and failing) to find their current
directory entries.

Fixing this exposed an issue with the StateChecker: if a node is
added, modified and removed within a batch, there may be some ops that
have "updates" for BlockPointers that only existed during the batch,
and thus will never make it to the BlockPointer.  It would be better
to somehow edit the list of ops to not reference these blocks, but
that's fairly tricky given that nodes within removed directories could
have moved out before the parent was deleted, and we still need a
record of the original create somewhere.  The self-updates are pretty
harmless, so for now just ignore them in the StateChecker so that
tests will pass.

Issue: KBFS-2243