key_manager_test: fix tests now that some rekey events are ignored
Now that the rekey FSM ignores rekey requests when the promptPaper
flag is already set, it caused some tests to wait for long timeouts
for the events they sent to simulate the server returning their own
rekey-bit-set update.  So instead, don't wait for a complete rekey
event to finish in that case, just wait for the simulation event to be
processed.