tlf_journal: enable single-team journals with team ID
In a future PR, the `tlfJournal` will need to know the team ID
corresponding to a single-team TLF, so that it can check the team's
quota when limiting space, rather than the user's quota.
Unfortunately, getting the team ID here is not very convenient.

This PR plumbs it through when the journal is created in the first
place, if possible.  If not possible, it trusts the MD server to
provide the handle, given a TLF ID.  It saves the team ID to the
journal info file on disk, so it can be used without needing server
access after a restart.

Issue: KBFS-2217