journal_md_ops: in GetForHandle, don't cache server's update
We only need the server update to learn the TLF ID.  If there is
already a local revision in our journal and cache, we don't want to
override it.  We also don't need to do the full processing, we only
need to check that the ID has been signed over.

Issue: KBFS-2310