mdcache: use Replace for local puts, and use a lock in mdcache
This is to avoid a race where we could fetch something from the server
and put it in the cache, all while the local folderBranchOps is
creating a new MD with the same revision number, which will then get
ignored when its tried to Put into the cache.

Instead, use Replace for locally-created MDs, and use a lock in
MDCache to get rid of this race completely.

Issue: KBFS-2310