vendor: update go-git to master and add push optimizations
Compared to the v4 branch we were using before, master branch seems to
fix a few bugs and bring the time for pushing the KBFS repo from hours
to about 25 minutes.  With the following optimizations, I got that
time down to less than 3 minutes.

* src-d/go-git#564
* src-d/go-git#565

I suspect there are still lots more optimizations that can be done.  I
don't think there are many users of Push in this project.