libfuse: serve FS requests before the mounter is ready to fix hang
In some cases, /keybase was not being mounted for macOS users,
although the kbfsfuse binary was running fine.  On investigation, it
looked like the osxfuse_mount call was giving up silently after 2
minutes.

According to osxfuse developer Ben Fleischer:

> This is most likely the reason for the issue you are seeing. In rare
> cases mount(2) performs a FUSE_STATFS operation after mounting the
> volume, but before returning control to mount_osxfuse. In this case
> calling mount_osxfuse synchronously will result a file system
> deadlock. I've ran into the same issue some time ago. To fix this
> mount_osxfuse needs to be called asynchronously. See
> https://github.com/osxfuse/fuse/commit/13ee6ec98e51cb76e2a6cfe80993e74c6073fe87
> for the relevant libfuse commit.

Instead, we'll wait for the mount to be ready in parallel with serving
new FS requests, just in case we need to respond to a STATFS call
early.  If the mount fails, we cancel the serving context (and
presumably the fd associated with the FUSE connection will be shut
down as well), which should cause the `startMounting` function to
return.

Issue: KBFS-2409