tlf_journal_test: sidestep signal race in singleOp test
When the background worker converts the MD journal to a branch, it
sends out a pause signal before it finishes and sends a `nil` error
back to the background loop.  Most of the time the background loop
processes the pause first, but sometimes it apparently sees the `nil`
first and goes into idle mode before going into pause mode.

I didn't see a way to force the loop to process one or the other
first, so I just made it possible to expect a set of signals rather
than just a single one.

Issue: KBFS-2440