kbfsgit: onDemandStorer shouldn't hide missing obj error
go-git depends on getting on error from `EncodedObject()` when an
object is missing, especially when constructing the set of "ignored"
objects.  This is the set of objects that the local git repo already
has. and there's no need for them to exist in the kbfs git repo.
go-git allows objects to be missing when constructing that ignored
list, but we were hiding that error in onDemandObject.cache().
Instead, we should surface it during `EncodedObject`, but without
fully reading all object data from disk, which would negate the point
of previous optimizations.

Issue: KBFS-2445