tlf_journal: don't use lock context on initial MD flush
Normally the lock context is used in single-op mode, and there will
always only be one revision to flush.  The exception is when the first
revision of the TLF hasn't been pushed yet, since it can't be
squashed.  In that case, there could be 2 revisions in the MD journal.
Don't unlock on the first flush, since we'll still need to hold the
lock until the second flush.