kbfsgit: lock during repo creation and on packed-refs creation
We need to take a lock before creating the repo, when generating the
repo ID, so there are no conflicts.  Use ._kbfs_config as the lockfile.

Also, any time we're pushing to a repo we think is empty, we need to
create a packed-refs file, even if it's only a single branch, so that
we can take a lock on the packed-refs file and ensure the repo only
gets that initial push once.

Finally, it turns out that calling WaitForBlockFlush() isn't enough
when in single-op mode, because the subsequent call to
SyncFromServerForTesting won't bother fetching anything from the
server when the local journal still has data in it.  So we have to
finish a complete single op, flushing all object files, before taking
a lock.

Issue: KBFS-2477