vendor: update go-git, so we don't over-transfer on a push
If the user is doing a push to a branch in the KBFS repo, and the KBFS
repo has extra commits not known to the local repo (i.e.., the local
user has forked), then go-git was causing way too much data to be
transferred because it doesn't fully negotiate the common commits
between the two repos.  It simply sends the commits representing all
references known by the KBFS repo; the local side wouldn't recognize
the new commit from the KBFS branch, and would therefore conclude the
user needs to push all commits in its history.  (Note: if the KBFS
repo has multiple branches, and at least one of which has a reference
with a common commit, then the entire history isn't sent, only the
commits up to that common commit, which is still wasteful.)

The right way to improve this is to implement the multi-round git ack
protocol (https://github.com/git/git/blob/master/Documentation/technical/pack-protocol.txt),
which keeps iterating until the two sides agree on a set of common
commits.  Unfortunately, go-git doesn't implement this, and it looks
like it would be a ton of work to implement it right.

Instead, do something quick-and-dirty in go-git.  Instead of sending
just the HEAD commit for each KBFS references, send the most recent
100 commits.  Hopefully, the local side will know at least one of
these, and can prune the set of pushed objects appropriately.  As an
optimization, we don't need to send the full 100 if we already know
for sure that the local side has the reference.  In practice, this
shouldn't add much latency or extra communication overhead, especially
in the common case when most references match between the two repos.

In the case where the local branch is ahead of the KBFS branch (i.e.,
normally the case for a push), this also shouldn't cause much extra
data to be sent, because almost certainly the local repo would have a
"remotes/origin" reference for the branch it was based on, so the
commit walker would be able to stop early once it runs into that
common commit.

Issue: KBFS-2486