folder_update_prepper: update new MD's root ID if needed
A user in the wild had this happen:
* Device A: Removed a bunch of files
* Device B: Removed a bunch of files
* Device A: Flush to server.
* Device B: Notice conflict, resolve to local journal.
* Device B: Connectivity failure, flush pauses.
* Device A: Perform QR.
* Device B: Re-run conflict resolution on top of QR.
* Device B: Flush new MD to server.

Unfortunately, the new MD pushed in the last step still had the root
block pointer from the QR's MD.  Basically, the folder_update_prepper
didn't do anything interesting in the final CR because the real
conflicts had already been resolved in the previous MD, and as a
result, it never updated the root block pointer to what it had
previously computed.

To get around this, if we pull in an update from a previous unmerged
resolutionOp, update the root block pointer on the new MD if it
matches that pointer.

Also introduce a test that reproduces the issue without this fix.

Issue: keybase/client#8832