tlf_journal: don't resume if we haven't paused yet
If the TLF journal converts a branch for squashing while its flushing
blocks, it doesn't yet signal for a pause, because that cancels the
existing context and would fail all the currently-flushing blocks for
no good reason.  However, if the branch then gets resolved before the
flushes finish, resume will get called before the pause happens on the
next loop, and there will be a panic.

Avoid that by only resuming if we've definitely paused.

Issue: KBFS-2502
Issue: keybase/client#8912