libgit: always set namespace for returned FS, and refactor a bit
We weren't correctly setting the repo ID as the lock namespace when
returning an FS for a newly-created.  That would only affect the first
push of public and multi-user TLFs that are lazily-created.

Also, factor out the taking of the lock config file into a separate
function for clarity, and reuse in a later PR.  Verify that the lock
namespace is what we expect it to be, to make sure a bad FS isn't
passed in by accident.

Issue: KBFS-2439