cr: don't fail if a direntry can't be found in identifyType
If this sequence of events happens on an unmerged branch:
* create foo
* setmtime foo
* rm foo

the crChain code would try to identify the type of foo because of the
setmtime operation, but it would fail because the directory entry for
foo no longer exists.  This is fine, since there can't be any
conflicts on it anyway.  The existing error was being overzealous.

Issue: keybase/client#9034