journal_server: disable journal auto-creation when logged out
If a user had viewed a team folder while logged in, and then later
logged out, the background quota reclamation process would try to
re-enable the journal for that team folder, but would cause a crash.
Instead, don't allow journal auto-creation when logged out.

This also re-arranges some error-checking in the journal enabling
code, to make a little more sense.  And it nixes the confusing error
message when `EnableExistingJournals()` is called multiple times in a
row with the same UID.

Issue: KBFS-2503