kbfsgit: check if GC is needed on every fetch/clone/push
For now, we just check if:
* GC hasn't been run on the repo in the last week.
* There are more than 50 loose refs.

If the check is taking more than a second, it prints "Checking repo
for inefficiencies... " to stderr as a progress indicator.

If GC is needed, it prints something like:

```
Tip: this repo could be sped up with some garbage collection. Run this command:
    keybase git gc repo

```

Issue: KBFS-2539