kbfs_ops: don't fetch unmerged MDs when in SingleOp mode
We never want single-op mode to start off with an Unmerged MD.  It
should never be served from the (fresh) journal, and there's no real
reason to check with the server for Unmerged MDs anymore.  If someone
happened to get themselves into a state where there's an Unmerged MD
on the server (because they disabled journaling and then hit a bug,
maybe?) then single-op operations will fail and we'll have to help
them debug it.  But that seems pretty unlikely.