libkbfs: populate the TlfHandle with a TLF ID, and resolve i-team
When parsing a TLF handle, we now first check whether there's a
matching implicit team with a TLF ID populated.  If so, we use that
handle.  Right now this will never happen, since the service doesn't
return a TLF ID for any implicit team.

For all other handles, we look up the TLF ID as part of the
handle-making process, and put it in the handle.

This adds at least one extra RTT to the servers during handle parsing,
which is unfortunate while e.g. listing favorites, but it's necessary
for the implicit team design.

Issue: KBFS-2598