kbfsmd: add a new ImplicitTeamsVer (MDv4!)
This adds a new metadata version (4) that is data-compatible with v3
and uses the same struct.  It only uses the new version for
i-team-backed TLFs, otherwise it continues to use v3.  For now, prod
still uses v3 no matter what, until we get a little more testing done.

Issue: KBFS-2621