folder_branch_ops: enable journal when setting first TLF head
Once iteams are enabled and TLF IDs are stored in the team's sigchain,
`KBFSOpsStandard` will be able to create a `folderBranchOps` without
ever passing the TLF's handle through `journalMDOps`.  This is a
problem, because the handle is required to enable a journal.  Without
this, the journal won't get created until the first MD put with a
revision > 1, which is too late because that will happen when the TLF
is locally dirty, and we can't enable journals then.

So instead, enable the journal (if needed) the first time a head is
set inside the FBO.

Issue: KBFS-2621