libgit: implement a Clone function
This function "clones" a repo from a billy filesystem, into another
billy filesystem.  However, it doesn't make a true git clone, for two
reasons.

  1. This will be used to clone KBFS repos, and go-git doesn't
natively understand the keybase:// protocol.  There might be a way
around this with the `InstallProtocol` go-git method, but I haven't
investigated that yet and it would be some extra work.
  2. These clones are intended for autogit, which will display them as
read-only directories.  They do not need to be real git repos, and so
there's no need to copy object files or much else in the .git repo.
All we really need is an index, to compute the differences.  Making a
full git clone, or even a "shallow" git clone, would be wasteful in
that it would read and write way more data than we really need.

To make this happen, we open the existing bare repo, but fake out the
config in such a way that it stores its index, and any reference
modifications, in a .git/ directory in the worktree, but everything
else goes through the main bare repo's .git directory.  Then we can
just use a simple `Reset` call to set up the cloned files.

Issue: KBFS-2672