libgit: change Clone() to Reset() and add tests
Turns out that, from the perspective of the worktree, there's no real
difference between clone and pull for autogit, now that we're just
doing a hard reset to do the checkout.  So rename and add tests.

A later PR will still distinguish between clone and pull at a higher
level, within the autogit manager.

Issue: KBFS-2673